Laravel From Scratch [Part 11] - Access Control

+++ app/Http/Controllers/PostsController.php
@@ -11,6 +11,16 @@ use DB; // Чтобы использвовать обычный SQL запрос
 class PostsController extends Controller
 {
     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index','show']]);
+    }
+
+    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
@@ -79,6 +89,12 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        // Check for correct user
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error', "Unauthorized Page");
+        }
+
         return view('posts.edit')->with('post', $post);
     }
 
@@ -98,6 +114,12 @@ class PostsController extends Controller
         
         // Create Post
         $post = Post::find($id);
+
+        // Check for correct user
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error', "Unauthorized Page");
+        }
+        
         $post->title = $request->input('title');
         $post->body = $request->input('body');
         $post->save();
@@ -113,6 +135,12 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        // Check for correct user
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error', "Unauthorized Page");
+        }
+
         $post->delete();
         return redirect('/posts')->with('success', "Post removed");
     }
